pyproject.toml
[project]
name = "pdf_extractor_upgradevip"
version = "1.0.0"
dependencies = [
    "docling>=2.63.0",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "pypdf>=6.4.0",
    "reportlab>=4.4.5",
    "streamlit>=1.51.0",
]
